Reindex notes after a migration if needed
