+++ cmd/index.go
@@ -2,13 +2,6 @@ package cmd
 
 import (
 	"fmt"
-	"os"
-	"time"
-
-	"github.com/mickael-menu/zk/adapter/sqlite"
-	"github.com/mickael-menu/zk/core/note"
-	"github.com/mickael-menu/zk/util/paths"
-	"github.com/schollz/progressbar/v3"
 )
 
 // Index indexes the content of all the notes in the notebook.
@@ -18,7 +11,7 @@ type Index struct {
 }
 
 func (cmd *Index) Help() string {
-	return "You usually don't need to run `zk index` manually, as notes are indexed automatically before each zk invocation."
+	return "You usually do not need to run `zk index` manually, as notes are indexed automatically when needed."
 }
 
 func (cmd *Index) Run(container *Container) error {
@@ -27,33 +20,10 @@ func (cmd *Index) Run(container *Container) error {
 		return err
 	}
 
-	db, err := container.Database(zk.DBPath())
+	_, stats, err := container.Database(zk, cmd.Force)
 	if err != nil {
 		return err
 	}
-
-	var bar = progressbar.NewOptions(-1,
-		progressbar.OptionSetWriter(os.Stderr),
-		progressbar.OptionThrottle(100*time.Millisecond),
-		progressbar.OptionSpinnerType(14),
-	)
-
-	var stats note.IndexingStats
-	err = db.WithTransaction(func(tx sqlite.Transaction) error {
-		stats, err = note.Index(
-			zk,
-			cmd.Force,
-			container.Parser(),
-			container.NoteIndexer(tx),
-			container.Logger,
-			func(change paths.DiffChange) {
-				bar.Add(1)
-				bar.Describe(change.String())
-			},
-		)
-		return err
-	})
-	bar.Clear()
 
 	if err == nil && !cmd.Quiet {
 		fmt.Println(stats)
